support-figma: add allowlisted domains for Figma plugins
This change defines the `networkAccess` field in the Figma plugin and
widget's `manifest.json` files. This will prevent cross-origin
requests to domains not matched by any item in the `allowedDomains` field and
will update their Community pages to reflect that network access
restrictions are in place (as opposed to "Unknown network access").

There are currently only two cross-origin requests being made. The first
fetches the [figma-plugin-ds](https://github.com/thomas-lowry/figma-plugin-ds)
stylesheet and the other(s) are made by the stylesheet in an attempt to load
some Inter fonts from rsms.me using `@font-face`.

Figma pre-loads Inter for plugins to use making it unnecessary to fetch them
again so I've intentionally excluded those domains from the allow-list. There
are a few errors that show up in the dev console due to the CSP blocking the
requests to fetch the fonts, but those can be safely ignored[^1].

I also added a section in the README's that calls out the network access
restrictions.

Fixes google/automotive-design-compose#65

[^1]: Technically only the Regular and Bold styles are being pre-loaded,
      but Figma appends `https://static.figma.com/webfont/` to the
      `font-src` directive in the CSP and the rest of the font family
      can be fetched from there if need be. This isn't officially
      documented anywhere as far as I can tell so it's probably better
      to just expand the allowlist if other fonts are needed.