squoosh: fix a bug where animations to the base state could get lost when interrupted.
This was due to "null confusion". Null is a valid target node name, but
can't be used as a key in a HashMap, so we have a constant string
"<<null>>" that we use instead. We had one comparison where "null !=
'<<null>>'" which resulted in losing track of an animated transition to
the "null" target node once it had been started.

The "Smart Animate: Enum" test case now always correctly animates the
red/white component with dots, where previously the animation back to
the base/null state was lost when the counter updated.