squoosh: remove failed interaction transitions immediately, fix a layout id bug.
We now tell the interaction state if we were unable to run an animated
state transition, so it can move the state to whatever the completed
value was immediately.

There was also a bug where layout ids would get reused across documents,
even though the root layout id was different (the root layout id was not
being used for some component instances). This is fixed, and now
switching between "SA: Basics", "SA: Instances", and back doesn't result
in an empty "SA: Instances" page.