Closes #601: Fix a crash when a rotation dial/gauge is set to "Scale"… (#602)
… constraints.

When a node's constraints are set to "Scale", the ViewStyle's width and
height fields are Auto, causing the rendering code to be unable to get
the node's with and height. This causes the transform being used to be
full of NaNs, and then ultimately causes a crash.

To fix this, save the original node's size into ViewStyle into the
node_size field, renamed from text_style, and use it when the style's
width/height are not known.

Note that this does not fix support constraints on nodes -- it just
prevents a crash and allows the rotation to still function given that it
stays the same size as in Figma.