Closes #630: Rotated nodes don't show linear gradients properly (#639)
When getting the brush to render the background of a node, use same size
as the one used when rendering the node. Consolidate the code that
calculates this size into a function in Utils. This ensures that the
linear gradient parameters specified in Figma match the node size. This
also fixes an issue with angular gradients that regressed in the "Vector
rendering" test, and can be seen in the new render example image.

Update serialized file version, dcf files, and images of tests that have
changed.