squoosh: initial take at overlays.
This is a very basic attempt at implementing overlays in squoosh; it's
not fully complete and seems to expose some issues with interactions on
child composables (perhaps because the interaction coroutine is
destroyed when a child composable is removed from the tree, which
happens often for overlays).

Needs more debugging work, but it seems like overlays will fit the
design OK. Hopefully we'll be able to use the existing Squoosh animation
framework to implement the overlay transitions.