squoosh: fix text layout invalidation and add a test.
We keep a cache of objects that have been sent to Rust layout, and only
send them again when they change. We weren't consulting the text info
when calculating the cache key, so updated customized text didn't result
in a layout invalidation.

I added a custom hashCode implementation to Squoosh's layout data type
to ensure that we don't hash Compose's font loader, and don't hash more
than we need to.