squoosh: compute layout in Compose's layout phase, and consult incoming constraints.
This provides some basic integration with Compose's layout system, so
`Modifier.fillMaxSize()` works on a Squoosh-backed DesignCompose root
component now.

We can also do a better integration of layout with Composables inserted
into a DesignCompose-generated Component by propagating their intrinsic
size to the Rust layout implementation.

One further step would be to use a Rust layout "layout func" to measure
child Composables... but this would potentially break the "single pass
rule" of Compose layout, because the flexbox algorithm can try to
measure children multiple times with different constraints in some
circumstances. I had thought this might be an advantage of the Squoosh
approach (that we could do layout func integration), but in practice it
feels like something we should not do, and instead just document the
limitations.