Restore UntrackedTask annotations on Cargo tasks (#701)
Gradle tries to optimize build times by detecting when tasks do not need
to be re-run. This works well for many tasks, but gets tricker for more
complex ones, such running `Cargo build`. While you can add
configuration to inform Gradle of task inputs and outputs, so that it
can decide whether a task is up to date or not, this is very difficult
to do for a Cargo build, as there isn't a known mechanism for detecting
all of the inputs that Cargo would use to decide what to rebuild. Also,
Cargo already has it's own up-to-date checks and they generally complete
within a few milliseconds.

The `@UntrackedTask` annotation informs Gradle to not try to track
whether the task is up-to-date and can be skipped, and instead tells it
to always run the task. The annotation had been applied to the original
CargoBuildTask, but that task was refactored into the
`CargoBuildHostTask` and `CargoBuildAndroidTask`. Apparently the
annotation doesn't apply to descendants of the base class. Applying the
annotation to the new tasks fixes the below issue.

Fixes #697