Closes #705: Add a function to enable overriding document IDs
Add a function DesignDocOverride(id) that can be used to override the document ID used by any child/descendent generated DesignCompose functions. DesignDocOverride internally just uses a CompositionLocalProvider to provide the document ID, which is checked and used in the DesignDocInternal function.

Add a HelloBye example that uses this function to change the document ID.