Closes #397: Implement scrolling support for autolayout containers
Use Modifier.scrollable() to enable scrolling on autolayout frames. Handle the scrolling logic manually by keeping track of the scroll offset and passing it to layout. The layout function also sets the extents of the scrollable frame's children in order to establish the scrollable bounds.

Note that scrolling to the end of a list does not have the same overscroll bounce effect that native scrollable containers have.

This commit also re-enables the DesignSwitcher since with the scrolling change you can scroll through log messages again.