squoosh: always pass the variant details to the variant transition manager.
This ensures we don't need to do any null handling in
SquooshVariantTransition, and makes it possible to do some very basic
filtering of which variant transitions get animated there.

Slightly longer term, we need a way to control variant transitions --
what gets an animation, and the properties of the animation.