Update layout code to use placeWithLayer() instead of place()
Cleaned up unused fields and data structures for text measurement
Optimized text remeasurement by only resubscribing to layout if data in the text changes. Previously, it was resubscribing too often because ParagraphIntrinsics is not a data class and its equality comparison would return false, triggering the DisposableEffect that subscribes to layout to run unnecessarily.