Closes #397: Implement scrolling support for autolayout containers (#720)
Use Modifier.scrollable() to enable scrolling on autolayout frames.
Handle the scrolling logic manually by keeping track of the scroll
offset and passing it to layout. The layout function also sets the
extents of the scrollable frame's children in order to establish the
scrollable bounds.

Note that scrolling to the end of a list does not have the same
overscroll bounce effect that native scrollable containers have.

Cleaned up unused fields and data structures for text measurement

Optimized text remeasurement by only resubscribing to layout if data in
the text changes. Previously, it was resubscribing too often because
ParagraphIntrinsics is not a data class and its equality comparison
would return false, triggering the DisposableEffect that subscribes to
layout to run unnecessarily.