Add missing git implementation selection to vergen cargo package and update to 8.3.1
I'm not sure how long this has been required, but according to [vergen's docs](https://docs.rs/vergen/latest/vergen/#configuring-the-git-feature):

`If you wish to use the git feature, you must also enable one of the git implementations. The gitcl features is lightweight, but depends on git being on the path`

Dependabot is trying to update vergen to 8.3.1 in #737 and it's failing to build. Adding the git implementation seems to fix this.

I've selected the `gitcl` implementation because it's the most lightweight. It requires `git` to be in the build system's path, but I don't think that will present any problems. If it does then we can change to one of the other implementations that doesn't depend on `git` being available.