Replace KTFmt Gradle Plugin with Spotless.
I've been having consistent problems with the ktfmt Gradle plugin and
this time it blocked me from updating to Gradle 8.6. I've been meaning
to replace it with Spotless for a while, which can be expanded to also
handle licenses, markdown, and other formatters. Spotless seems better
supported in general and has allowed me to remove most of the custom
configuration that I had to add to the ktfmt Gradle plugin to get it to
also format our build scripts.