Closes #843: Remove the requirement to propagate ParentLayoutInfo for replacement data
Before this changes, implementing a component replacement or content replacement customization required the developer to pass a ParentLayoutInfo object from the original node through to the replacement node. This allowed the replacement node to use the original node's layout and size properties. This resulted in boilerplate code with every replacement.

This change removes this requirement and instead moves ParentLayoutInfo from a parameter into a CompositionLocalProvider. This allows the correct ParentLayoutInfo object to be declared by the caller of the replacement component composable so that when that composable is called, it has the correct ParentLayoutInfo in scope.