Closes #864: Fix possible crash when using component replacement with a non-DesignCompose generated function
When using a component replacement customization with a non-DesignCompose composable, the replacement composable would not subscribe to layout changes. If the original component had any siblings, those siblings would subscribe to layout with a missing child index, causing a crash in Rust layout when trying to insert a style node at an invalid position. For example if nodes A, B, and C were siblings and B was the replacement node, A and C would subscribe to layout with indicies 0 and 2 respectively, and Rust would fail trying to insert a style node at index 2 when the list was only had a lenght of 1.

This change fixes the crash by detecting when a component replacement does not call a DesignCompose composable, and in that scenario will subscribe to layout anyway with the original node's style.

Secondly, the Modifiers passed into ComponentReplacementContext have been changed. The appearance modifier has been removed, and the layout modifier has been changed to be a modifier that holds layout data for the original node. When using DesignCompose composables as replacement content, this modifier is not needed, but when using non-DesignCompose composables, the layout modifier should be passed into the composable in order to use the original node's size and position.