squoosh: introduce an (optional, disabled) tree management strategy. (#859)
Squoosh introduces a new approach for managing the view tree to
DesignCompose. With the introduction of f/rustlayout, we actually don't
need to make a Composable for every node in the design asset, and can
instead only emit Composables when we need interactions or
accessibility.

Because we consolidate management of the tree of design elements, it's a
bit easier to perform layout animations (as we can simply have two
trees, apply the relevant state to both, do layout on both, and then
create a combined tree that allows transitioning). When Squoosh is
enabled some basic "Smart Animate" functionality works.