Closes #916: Support customization flattening into modules (#906)
Support a new @DesignModuleClass annotation to declare a class that
holds customizations. These classes can be used to consolidate
customizations into a single parameter in a @DesignComponent function,
as well as nested in another @DesignModuleClass.

When declaring a @DesignModuleClass, use the @DesignProperty and
@DesignVariantProperty annotations instead of the normal @Design and
@DesignVariant properties. This is needed because of a bug in KSP that
prevents annotations from being used with multiple targets. Once fixed,
we can remove @DesignProperty and @DesignVariantProperty.

A @DesignModuleClass parameter can be specified in a @DesignComponent
function with the @DesignModule annotation. A @DesignModuleClass class
property can be specified in a @DesignModuleClass class with the
@DesignModuleProperty annotation. Once the KSP bug is fixed, we can
remove the @DesignModuleProperty annotation.

This change creates a new source file BuildModuleClass.kt to process
@DesignModuleClass classes. Each @DesignModuleClass will generate a new
file associated with that class that has an extension function for the
class. The extension function creates a CustomizationContext object,
populates customizations for it, and then returns it. Node queries and
ignored images are also tracked so that any @DesignComponent function
that uses modules will generate the appropriate queries() and
ignoredImages() functions.

This change also cleans up some of the code in BuildProcessor.kt and
moves common functionality into a separate file.