Fix #949: use Figma animation duration/curve settings. (#972)
Update DesignCompose's serialization format to support spring-based
animations (as well as cubic bezier + duration). This representation
fully supports all of the options exposed by Figma.

Now, for animated interactions, the transition values set in Figma will
be used. For variant/enum-based transitions, a default spring value is
supplied. A future change will make it so that this value can be
customized, too.