Fix #1052: rewrite squoosh animation tree building to not mutate trees.
While attempting to fix a bug in squoosh's animation support where the
root frames of components were not considered for animation, I kept
running into layout assertions, which ultimately were because the
animation implementation performs destructive mutations on the base
layout tree AND on the "transition layout tree".

This change means that all `SquooshResolvedNode` trees are now immutable
after construction (except for layout and style values which can change
after the tree was built).

Finally, because the source trees are immutable, we can run layout on
the root and target root trees whenever layout needs to change. This
means animations are uninterrupted by relayouts, and we can relayout as
many times as we like without generating new trees.

The test expectation for the animation midpoints test is updated because
the root node of the small variant test actually changes itself in a way
that we don't support via interpolation -- so we can't recurse into the
children, we can only crossfade. This behavior doesn't match Figma,
which is able to do interpolations of simple fills and strokes.