Closes #948: Remove support for interface inheritance (#1007)
With design modules available, there should be no need for interface
inheritance in @DesignDoc interfaces. Remove support for them and show
an error if inheritance is attempted. In order to support upgrading the
Compose compiler, which doesn't support default args in open functions,
change the generated code to create a class instead of an interface.