Refactor the layout and figma_import crates (#1006)
As a result of this change the `figma_import` crate now depends on the
`layout` crate, instead of the `layout` crate depending on
`figma_import`.

As a result, the `layout` crate now has no dependencies within the
workspace.

`layout`'s tests had to be moved to `figma_import` due to the tests
heavy reliance on `figma_import`.