Migrate the Layout Nodes structs to the Layout crate (#1074)
While working on the Protobuf code I realized that we never finished
changing the LayoutNode structs that are sent across the JNI to use
LayoutStyle instead of the full ViewStyle. In order for me to finish the
first set of Protobuf changes I'll need to get these moved to the proper
crate.