Closes #1056: Fix view finding for @DesignComponent functions on a component set.
Historically when a @DesignComponent is declared with a component set, the component set name must match the property name in a @DesignVariant parameter in order for the view matching algorithm to use the correct variant at runtime. This has always been confusing and undocumented, so this change allows the property name to differ from the component set name. The view searching algorithm now checks at runtime of a node is actually a variant of a component set, and uses that if so. Then, in order to support TapCallback using the name of the component set, populate variantParentName if we have the component set info to do so.

The VariantProperties test has been updated to test this scenario.