Closes #1141: Fix crash with 0 width/height vectors
When a vector has a width or height of 0, is set to scale constraints, and has a parent node, some of its layout properties were getting set to invalid numbers due to a divide by zero. This change fixes this bug by setting the values to 0 instead. A unit test has been added for this scenario.