Add default Android Studio Settings (#1146)
<!-- start git-machete generated -->

# Based on PR #1142

## Full chain of PRs as of 2024-05-24

* PR #1146: `wb/froeht/android_studio_settings` ➔
`wb/froeht/vscode-settings`
* PR #1142: `wb/froeht/vscode-settings` ➔ `main`

<!-- end git-machete generated -->
Please give this a shot and let me know what you think.

This **WILL** conflict with your existing files. These are the files
that GitHub and IntelliJ recommend you check into version control. (See
the additions to .gitignore)

I've checked each file to make sure that the settings are appropriate
and enforce our style.

The formatting defined here should NOT conflict with the formatting
options defined in the VS Code config that's been committed. If it does
then try to figure it out and upload a fix.

This includes proper formatting rules for yaml and markdown files, so
you'll see those get updated if you modify them (such as the
github-pages.yml file that I used to make sure Android Studio and VSCode
were formatting it the same)

As a bonus, I've also added a bit of configuration to have Android
Studio automatically run the generate sources tasks (KSP and Serdegen)
when the project opens. This will give you a head start.

Also, I found the setting to make Android Studio use Git's file staging
by default, so now it and VS Code's source control views will work the
same!