Set inter font for all validation tests (#1223)
This should fix issue #1186 that github workflow generates consistent
screenshot diffs for variant tests. The verifyRoborazzi check are not
failing any more.

I believe this is similar to the interaction states not get cleared. On
the local machine, it seems tests from different kt files are running on
different processes but the github flow somewhat runs different kt test
files on the same process.

Right now some of the tests have the font set and others not. So I am
setting font to all tests so the screenshot are generated using the same
font no matter if it shares a process with others or not.

But further more, we might want to look into why and how to isolate
tests?? I haven't found anything how to do that for Robolectric though.