Migrate the styles from layout to dc_bundle.
This is a small example of what will be done to complete #1241.

The enums were moved (including the protobuf conversion code from #1079)
which meant that the `Into<taffy::..>` impls were now orphaned.

This was fixed by creating a new `IntoTaffy` trait within the Layout crate.

I'm moving the "legacy" versions of the structs and enums into a "legacy_definition" module in dc_bundle. Once the protobuf transition is complete we'll be able to delete the module.