Towards #1213: improve component replacement implementation.
This change improves basic component replacement so that more of the
component replacement tests pass, and a baseline image for component
replacement with squoosh has been added.

There is no real connection in the layout trees in this implementation.
The parent decides the size of the children and the children get laid
out in that size. If a child needs more space, then it has no way of
requesting it.

We could make it so that the parent uses the Compose measure method
first, but it's not clear what we need to pass in as constraints (it
could be the style-based min/max width/height if available...).