Closes #917: Support rotations in squoosh animations
First, this change fixes rendering of rotated nodes in Squoosh by rendering with the original node size instead of the layout-calculated node size. To prevent animation regressions for nodes whose size changes, a new 'overrideLayoutSize' field has been added to SquooshResolvedNode. When this field is set to true, which is during a layout animation, the renderer always uses the layout sizes computed by the animation code. When this field is set to false, the renderer tries to use the original node size, if it was fixed.

To support rotation animations, we decompose the source and destination matrices, interpolate their values, and then construct a matrix with the interpolated values. Some rotation animations have been added to the smart animate test, and a midpoint animation test has been added for rotations.