Migrate the styles from layout to dc_bundle.  (#1248)
<!-- start git-machete generated -->

# Based on PR #1280

## Full chain of PRs as of 2024-06-21

* PR #1248: `wb/froeht/migrate-styles-to-bundle` ➔
`wb/froeht/fix-proto-package-names`
* PR #1280: `wb/froeht/fix-proto-package-names` ➔ `main`

<!-- end git-machete generated -->

This is a small example of what will be done to complete
https://github.com/google/automotive-design-compose/issues/1241.

The enums were moved (including the protobuf conversion code from
https://github.com/google/automotive-design-compose/issues/1079)
which meant that the `Into<taffy::..>` impls were now orphaned.

This was fixed by creating a new `IntoTaffy` trait within the Layout
crate.

I'm moving the "legacy" versions of the structs and enums into a
"legacy_definition" module in dc_bundle. Once the protobuf transition is
complete we'll be able to delete the module.