Migrate FlexWrap to new proto message (#1210)
<!-- start git-machete generated -->

# Based on PR #1280

## Full chain of PRs as of 2024-06-21

* PR #1210: `wb/froeht/proto-impl-rust` ➔
`wb/froeht/fix-proto-package-names`
* PR #1280: `wb/froeht/fix-proto-package-names` ➔ `main`

<!-- end git-machete generated -->
Please read #1209 before reviewing.

This PR is a bit of a tester, replacing the native FlexWrap enum with
the one generated by the protobuf compiler.

As you can see I had to make a small change to the proto file to get it
to behave properly (in this case the enum was being generated with
"FLEX_WRAP_NO_WRAP" being converted to "FlexWrap::Nowrap" instead of
"FlexWrap::NoWrap").