Add plugin support for localization (#1171)
Add plugin support for localization, which goes through the text nodes
and generate a strings xml file out of it, where the string resource
name is created from either node characters or node name.
If there are style overrides, it will generate a string array.

Update DesignCompose to support string resource or string-array resource
for text nodes.
Add an option to DesignSwitcher so when it is unchecked, it will use the
text characters from the figma file and then live update still works.

Add a menu to clear the string resource name from the figma file.