Closes #917: Support rotations in squoosh animations (#1263)
First, this change fixes rendering of rotated nodes in Squoosh by
rendering with the original node size instead of the layout-calculated
node size. To prevent animation regressions for nodes whose size
changes, a new 'overrideLayoutSize' field has been added to
SquooshResolvedNode. When this field is set to true, which is during a
layout animation, the renderer always uses the layout sizes computed by
the animation code. When this field is set to false, the renderer tries
to use the original node size, if it was fixed.

To support rotation animations, we simply interpolate the transform
matrix.