Closes #1369: Support progress bars that move along a vector path
Vectors in Figma that have a stroke path, not a fill path, can now be used as a progress bar. The Dials and Gauges plugin now has a "Progress vector" section that can be enabled on vector nodes. Currently there is no preview feature for vector progress bars, as there is not a straightforward way to render a partial vector path using Figma's APIs.

DesignCompose applications parse this new vector progress data an render a partial path by using the built in dashed path effect.

This commit also addds tests for dials & gauges