Closes #1437: Support Round and Square stroke caps for vector progress bars
Add a new validation test with vector progress bars using various stroke backgrounds, with round and square end caps. Add a screenshot test for this as well. This also fixes a bug where turning a vector into a progress bar vector using the plugin and then turning it off would still render it as a progress bar because the "enabled" flag was ignored.