Add a group option which allows to use different string resource name for nodes with same text (#1445)
Fixes: #1371
Also made a change when user rename a string resource name to an
existing one, it allows to merge the two together.

This user action will not be saved to the plugin data so it gets lost
the next time running the plugin. This is also because user can use a
different group option which may change the string resource name
assigned to a text node.

---------

Co-authored-by: Robert Lin <130500397+rylin8@users.noreply.github.com>