Closes #1493: Support progress bar and marker constraints (#1496)
Support layout changes with progress bars and markers in squoosh. When
constraints are set on a progress node, use the parent layout to
calculate the extents of the progress bar/marker. This is only supported
on squoosh since we have access to the parent layout info there.
Additional tests have been added for this.