Fix #1576: Improve performance by caching text layouts and computed paths.
Profiling showed that computing text layouts and computing paths was
taking the bulk of runtime in a simple demo app. This change adds two
generational caches -- one for computed text layout (e.g.: the Compose
`ParagraphIntrinsics`), and one for computed paths (the `ComputedPath`
type which contains all of the paths needed to render a node).

Additionally some unused logging has been removed.

In a profile of "FutureV2" design, `squooshComputeTextInfo` run time was
reduced by a half, and `computePaths` was reduced by around 5x.