Fix #1503: ensure child composables are drawn, by using Layout ID rather than SquooshResolvedNode identity. (#1504)
The previous check (that the SquooshResolvedNodes were equal) failed,
because during a transition we generate a new set of
SquooshResolvedNodes from the source tree and dest tree. With this
change, we consult the layout ID, which is unique within a tree (so
still useful for this purpose of identifying the selected node for child
rendering).