Add plugin to export images (#1593)
Images are downloaded in png format for png and jpg. Gif images will
also be downloaded, in the gif format. But there is only one image hash
returned by the figma plugin api which is for the gif itself. The REST
API however returns an `imageRef` which might be the image hash of the
first frame and a `gifRef` which is the image hash returned by the
plugin api.

The plugin data is in {image_hash: res_name,...} format.

Right now all images are exported. Will allow user to pick which images
to export in a following pr.

Towards: #1556