Convert Dimension and DimensionRect to Proto (#1598)
Dimension and DimensionRect were a big pain, because Dimension is a
typed enum and proto doesn't handle those very well.

On the Rust side I basically preserved the original Dimension and
DimensionRect as proxies for the Protobuf versions. The proxies offer
much cleaner code than using the proto types directly and convert to and
from the proto types easily.

I did end up with a number of unrelated changes from running cargo fix,
which we should really be running more often. (I should add it to the
format script).

I didn't do the same on the Kotlin side because there is much less
manipulation of the types, but the same could be done as well.

The resulting code is uglier than I'd like, but we need to get this
finished and move on.

<!-- start git-machete generated -->

# Based on PR #1634

## Full chain of PRs as of 2024-09-19

* PR #1598:
`wb/froeht/1575-ProtoConv-Convert-Rectangle` ➔
`wb/froeht/fix-robo-test-recording`
* PR #1634:
  `wb/froeht/fix-robo-test-recording` ➔ `main`

<!-- end git-machete generated -->