Add support for animating text and vector sizes by scaling (#1680)
Fixes #1641. Set the scale value in the transformation matrix for text
nodes that animate to a different font size, and for vector paths that
change size. Do this with a crossfade so that it will work with other
changes such as font weight as well.

Modify the fade in/out animations to account for the target node
opacity.

Refactor the AnimatedLayoutItem into more functions so that we can more
easily support additional style animations in the future.