Fix item spacing of list replacement content in squoosh
Fixes #1541. Set the default flex_grow value to 0 for list replacement items. This prevents the items in the list from stretching to fit the container they are in.

This change puts several of the validation tests previously in the DEFAULT_RENDERER_ONLY category into the BOTH category. Squoosh is now able to render these examples correctly, with some (CrossAxis Fill) being more correct than the non-squoosh version.