Refactor: Update ViewShape to use proto enums
This change updates the ViewShape struct to use proto-defined enums for StrokeCap and Shape. This improves type safety and consistency across the codebase.

Specifically:

- Removes `legacy_definition::element::view_shape::ViewShape` and `legacy_definition::element::view_shape::StrokeCap`.
- Introduces `definition::element::view_shape::StrokeCap` and `definition::element::ViewShape`.
- Updates `figma_schema::StrokeCap` to convert to the new `definition::element::view_shape::StrokeCap`.
- Adds helper functions to `ViewShape` for creating different shape types.
- Replaces instances of the old `ViewShape` and `StrokeCap` with the new ones throughout the codebase.
- Updates `figma_import::reflection` to trace the new types.
- Updates the dcf files to reflect the changes.

In addition this required renaming ColorOrVar's inner type because both the inner and outer types shared the same name. This caused a collision in serdegen which caused it to error out.