Add new Cargo fetch tests and remove old mocked pseudo-test
This adds a new Cargo fetch test to check that a doc can be fetched and serialized. Because we don't want GitHub to be trying to fetch files, it's locked behind a new feature, "test_fetches".

This also removes the figma_v1_mock file and feature. This was a pseudo test that wasn't being executed by anything. It worked by using a feature flag to mock the http fetch, which meant that if the feature flag was included that no http queries would work, including the new fetch test.