Fix dcf_info to print info on older dcf files instead of showing an error.
Use a new header format that contains the document ID, name, last modified time, and Figma response version. Use this header to print out info when running dcf_info. To support older dcf files, try parsing the old header format, and if it is a version before this one (23), just print out the DC version and nothing else, since that's all the header provides and parsing the rest of the file would fail to deserialize.

Fixes #1790