Create fetch rule for live update (#1804)
Towards: #1028
Towards: #1801

When working on pr #1800 to create a separate shader test for
`CustomBrushTest`, I would like to disable it from the
`RenderAllExamples` test but then realized that it would not fetch the
dcf file with the gradle task. So I created the fetch rule and made the
`ShaderTest` and `StateCustomizationsUnitTest` fetchable in this pr.

Also explored a little bit further to use this fetch rule to make
`DesignSwitcherBasicTests` able to fetch dcf file.

Tested locally by running `./gradlew fetchAndUpdateFigmaFiles` and
`./gradlew recordRoborazziDebug`