Support customizations to get and set scroll position
Fixes #1812: Create a DesignScrollCallbacks customization where a developer can set two scroll callbacks. First, a setScrollableState callback that is called to get a ScrollableState object. This object can be used to set the scroll position. For grid views, this object is also a LazyGridState object that can be used to obtain additional information about the scroll state. Second, a scrollStateChanged callback that is called whenever the scroll state changes if the view is a row or column (not a grid). The callback provides the scroll position, max scroll position, frame size and content size.