Refactor: Inline EncodedImageMap into DesignComposeDefinition
This change refactors the way images are stored in the DesignComposeDefinition proto.
Instead of using separate EncodedImageMap and EncodedImage messages, the `images` field is now a direct map of image keys to byte arrays.
This simplification removes the need for extra message types and makes the definition more straightforward.

The existing EncodedImageMap has been promoted to the main definition file. It will be used in the code and then cloned to the Proto message before serialization.