Convert NodeQuery to Proto (#1837)
This one is a little more complicated because we use NodeQuery as they
key to a map in the DesignCompose Definition and protobuf messages can't
use a message as a key.

Fortunately NodeQuery only contains variations of a String, so it's
fairly easy to encode it to a string for use in the DCD. We'll still
keep the NodeQuery enum and just convert to/from it from the Protobuf
message.

Fixes #1670
Fixes #1282